pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "E-Campus Project"
authors = ["Tarlan Soltanov <tarlansoltanov@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
django-split-settings = "^1.2.0"
python-decouple = "^3.8"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.1"
django-filter = "^23.4"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
autoflake = "^2.2.1"
flake8-pyproject = "^1.2.3"

[tool.black]
line-length = 119
target-version = ['py39']
extend-exclude = '''
/(
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 119
py_version = 39
order_by_type = false
skip = [".git", "__pycache__", ".venv", "migrations"]

[tool.flake8]
max-line-length = 119
exclude = [".git", "__pycache__", ".venv", "migrations"]

[tool.autoflake]
check = true
recursive = true
remove-unused-variables = true
in-place = true
quiet = true
exclude = ["__init__.py"]
imports = ["django", "djangorestframework", "drf-yasg"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
